Exceção de négocio completa. Problema no relacionamento presente com loja.

+++ src/main/java/servico/BuffetServico.java
@@ -18,15 +18,13 @@ public class BuffetServico extends Servico
 
     public void salvar(Buffet buffet) throws ExcecaoNegocio
     {
-       
-
         if (existente(buffet.getValorTotalGasto()) == false)
         {
             em.persist(buffet);
+        } else
+        {
+            throw new ExcecaoNegocio(ExcecaoNegocio.OBJETO_EXISTENTE);
         }
-        else throw new ExcecaoNegocio(ExcecaoNegocio.OBJETO_EXISTENTE);
-
-        
     }
 
     public List<Buffet> listar()
@@ -34,30 +32,41 @@ public class BuffetServico extends Servico
         em.flush();
         return em.createQuery("select t from Buffet AS t", Buffet.class).getResultList();
     }
-    
+
     public List<Buffet> listarComComesBebes()
     {
-        em.flush(); 
+        em.flush();
         return em.createQuery("select t from Buffet AS t JOIN FETCH t.comesBebes", Buffet.class).getResultList();
     }
 
-
-    public void remover(Buffet buffet)
+    public void remover(Buffet buffet) throws ExcecaoNegocio
     {
-        Buffet t = (Buffet) em.find(Buffet.class, buffet.getId()); //se n tiver isso, o jpa acha que n deatachou        
-        em.remove(t);
+        /* CERIMONIA E BUFFET, SAO DOMINANTES. E SE RELACIONAM. 
+            A CERIMONIA EH MAIS IMPORTANTE QUE O BUFFET, LOGO CERIMONIA EXCLUE BUFFET.
+            O BUFFET NÃO PODE DELETAR CERIMONIA. POR ISSO ESSA EXCECAO EH LANÇADA AQ.        
+        */
+        if (buffet.associado() == false)
+        {
+            Buffet t = (Buffet) em.find(Buffet.class, buffet.getId());       
+            em.remove(t);
+        } else
+        {
+            throw new ExcecaoNegocio(ExcecaoNegocio.BUFFET_ASSOCIADO);
+        }
     }
 
     public void atualizar(Buffet buffet) throws ExcecaoNegocio
     {
-         em.flush();
+        em.flush();
 
         if (existente(buffet.getValorTotalGasto()) == false)
         {
             em.merge(buffet);
+        } else
+        {
+            throw new ExcecaoNegocio(ExcecaoNegocio.OBJETO_EXISTENTE);
         }
-        else throw new ExcecaoNegocio(ExcecaoNegocio.OBJETO_EXISTENTE);
-      
+
     }
 
     public Buffet buscar(int id)
@@ -67,7 +76,6 @@ public class BuffetServico extends Servico
 
     private boolean existente(double valor)
     {
-        System.out.println("servico.BuffetServico.existente()");
         TypedQuery<Buffet> query;
         query = em.createQuery("select b from Buffet b where b.valorTotalGasto = ?1", Buffet.class);
         query.setParameter(1, valor);
@@ -77,8 +85,6 @@ public class BuffetServico extends Servico
         {
             return false;
         }
-
         return true;
     }
-
 }
